[S10P12A306-657] feat: Add validation checks in admin Board api

See merge request s10-webmobile2-sub2/S10P12A306!133

+++ BackEnd/bid/bid-admin/src/main/java/com/ssafy/bid/domain/reward/service/RewardServiceImpl.java
@@ -9,17 +9,18 @@ import com.ssafy.bid.domain.notification.NotificationType;
 import com.ssafy.bid.domain.notification.dto.NotificationRequest;
 import com.ssafy.bid.domain.notification.service.NotificationService;
 import com.ssafy.bid.domain.reward.Reward;
+import com.ssafy.bid.domain.reward.dto.RewardListGetResponse;
 import com.ssafy.bid.domain.reward.dto.RewardSaveRequest;
 import com.ssafy.bid.domain.reward.dto.RewardSendRequest;
-import com.ssafy.bid.domain.reward.dto.RewardsFindResponse;
 import com.ssafy.bid.domain.reward.repository.RewardRepository;
 import com.ssafy.bid.domain.user.Student;
 import com.ssafy.bid.domain.user.repository.UserRepository;
+import com.ssafy.bid.global.error.exception.ResourceNotFoundException;
 
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
-@Transactional
+@Transactional(readOnly = true)
 @Service
 public class RewardServiceImpl implements RewardService {
 
@@ -28,43 +29,55 @@ public class RewardServiceImpl implements RewardService {
 	private final NotificationService notificationService;
 
 	@Override
+	@Transactional
 	public void saveReward(int gradeNo, RewardSaveRequest rewardSaveRequest) {
 		Reward reward = rewardSaveRequest.toEntity(gradeNo);
 		rewardRepository.save(reward);
 	}
 
 	@Override
-	@Transactional(readOnly = true)
-	public List<RewardsFindResponse> findRewards(int gradeNo) {
-		return rewardRepository.findRewards(gradeNo);
+	public List<RewardListGetResponse> getRewards(int gradeNo) {
+		return rewardRepository.findAllByGradeNo(gradeNo);
 	}
 
 	@Override
+	@Transactional
 	public void deleteReward(int rewardNo) {
+		boolean exists = rewardRepository.existsByNo(rewardNo);
+		if (!exists) {
+			throw new ResourceNotFoundException("삭제하려는 Reward 엔티티 없음.", rewardNo);
+		}
 		rewardRepository.deleteById(rewardNo);
 	}
 
 	@Override
+	@Transactional
 	public void sendReward(RewardSendRequest rewardSendRequest) {
 		List<Student> students = userRepository.findAllByIds(rewardSendRequest.getUsersNos());
 
 		if (students.size() != rewardSendRequest.getUsersNos().size()) {
-			throw new IllegalArgumentException("전송할 대상 중 확인되지 않은 사용자 존재"); //TODO: 커스텀 예외처리
+			throw new ResourceNotFoundException("Reward 전송 대상중 확인되지 않은 회원 존재.", rewardSendRequest.getUsersNos());
 		}
 
 		students.forEach(
 			student -> {
 				Reward reward = rewardRepository.findById(rewardSendRequest.getNo())
-					.orElseThrow(() -> new IllegalArgumentException("해당하는 리워드 존재 X"));//TODO: 커스텀 예외처리
+					.orElseThrow(
+						() -> new ResourceNotFoundException("전송하려는 Reward 엔티티 없음.", rewardSendRequest.getNo()));
 				student.addRewardPrice(reward.getPrice());
-				NotificationRequest notificationRequest = NotificationRequest.builder()
-					.receiverNo(student.getNo())
-					.title(reward.getName())
-					.content(rewardSendRequest.getComment())
-					.notificationType(NotificationType.REWARD)
-					.build();
+				NotificationRequest notificationRequest = createNotificationRequest(student, reward, rewardSendRequest);
 				notificationService.send(notificationRequest);
 			}
 		);
+	}
+
+	private NotificationRequest createNotificationRequest(Student student, Reward reward,
+		RewardSendRequest rewardSendRequest) {
+		return NotificationRequest.builder()
+			.receiverNo(student.getNo())
+			.title(reward.getName())
+			.content(rewardSendRequest.getComment())
+			.notificationType(NotificationType.REWARD)
+			.build();
 	}
 }
